split generic and upgrade specific context
